// FRONTEND/WorldCupOnline/src/app/MODELS/quinielaModel.ts
export interface quinielaModel {
    Id?: number;
    id_Usuario: number;
    id_Partido: number;
    id_Jugadores_goles_Eq1: number[];
    id_Jugadores_asistencias_Eq1: number[];
    id_Jugadores_goles_Eq2: number[];
    id_Jugadores_asistencias_Eq2: number[];
    Goles_Eq1: number;
    Goles_Eq2: number;
    Autogoles_eq1: number;
    Autogoles_eq2: number;
    id_Jugador_GOAT: number;
}